requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Pillow==7.2.0
SQLAlchemy==1.3.19
urllib3==1.25.9
WTForms==2.3.3
